backend/requirements.txt
Django==4.1.1
django-cors-headers==3.10.0
openai==0.27.1
halo==0.0.31
langchain==0.0.139
django-channels==0.7.0
channels_redis==4.1.0
daphne==4.0.0
djangorestframework==3.13.1
